+- SceneGraph Singleton & Unique Scene Actors
Changed the scenegraph into a singleton (should of done this sooner). There is still a bunch of pre-exisiting references to pre-singleton scenegraph that will need to be changed. Also made it so when you save a level it creates a folder and stores the actors in that folder. When you load a level it loads that exact folders actors, basically levels won't share the same actor XML files anymore.